Renamed package to Vehicles - Renamed url namespace to vehicles - renamed appconfig to vehicles
