+++ web/templates/navbar.html
@@ -20,8 +20,8 @@
         <li class="nav-item dropdown {% if vehicleActive %}active{% endif %}">
             <a class="nav-link dropdown-toggle" data-toggle="dropdown" href="#" role="button" aria-haspopup="true" aria-expanded="false">Fahrzeuge</a>
             <div class="dropdown-menu">
-                <a class="dropdown-item" href="{% url 'fahrzeuge:index' %}">Fahrzeuge anzeigen</a>
-                <a class="dropdown-item" href="{% url 'fahrzeuge:create' %}">Fahrzeug hinzufügen</a>
+                <a class="dropdown-item" href="{% url 'vehicles:index' %}">Fahrzeuge anzeigen</a>
+                <a class="dropdown-item" href="{% url 'vehicles:create' %}">Fahrzeug hinzufügen</a>
             </div>
         </li>
         </ul>
