+++ web/vehicles/tests/test_controllers.py
@@ -1,8 +1,8 @@
 from django.test import TestCase
 
-from fahrzeuge.controllers import VehicleController, VehicleTypeController
-from fahrzeuge.models import Fahrzeug
-from fahrzeuge.tests.test_models import VehicleTypeTestCase
+from vehicles.controllers import VehicleController, VehicleTypeController
+from vehicles.models import Fahrzeug
+from vehicles.tests.test_models import VehicleTypeTestCase
 
 
 class VehicleControllerTestCase(TestCase):
