+++ web/vehicles/urls.py
@@ -2,7 +2,7 @@ from django.urls import path
 
 from . import views
 
-app_name = 'fahrzeuge'
+app_name = 'vehicles'
 urlpatterns = [
     path('', views.IndexView.as_view(), name='index'),
     path('create', views.CreateView.as_view(), name='create'),
